@@ -11,14 +11,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^11.2.7",
-    "@angular/common": "^11.0.9",
-    "@angular/compiler": "^11.0.9",
-    "@angular/core": "^11.0.9",
-    "@angular/forms": "^11.0.9",
-    "@angular/platform-browser": "^11.0.9",
-    "@angular/platform-browser-dynamic": "^11.0.9",
-    "@angular/router": "^11.0.9",
+    "@angular/animations": "^11.2.8",
+    "@angular/common": "^11.2.8",
+    "@angular/compiler": "^11.2.8",
+    "@angular/core": "^11.2.8",
+    "@angular/forms": "^11.2.8",
+    "@angular/platform-browser": "^11.2.8",
+    "@angular/platform-browser-dynamic": "^11.2.8",
+    "@angular/router": "^11.2.8",
     "@fortawesome/angular-fontawesome": "^0.8.2",
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
     "@fortawesome/free-brands-svg-icons": "^5.15.3",
@@ -36,9 +36,9 @@
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.2",
-    "@angular/cli": "^11.0.7",
-    "@angular/compiler-cli": "^11.0.9",
+    "@angular-devkit/build-angular": "^0.1100.7",
+    "@angular/cli": "^11.2.7",
+    "@angular/compiler-cli": "^11.2.8",
     "@types/jasmine": "^3.6.9",
     "@types/node": "^12.20.7",
     "codelyzer": "^6.0.0",
@@ -52,6 +52,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "^4.0.7"
+    "typescript": "^4.1.5"
   }
 }
