@@ -1,10 +1,9 @@
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
+import { Routes, RouterModule, ExtraOptions } from '@angular/router';
 import { AboutMeComponent } from './about-me/about-me.component';
 import { ChangeLogAddComponent } from './change-log/change-log-add/change-log-add.component';
 import { ChangeLogComponent } from './change-log/change-log.component';
 import { FeatureAddComponent } from './change-log/feature-add/feature-add.component';
-import { ContactComponent } from './contact/contact.component';
 import { HomeComponent } from './home/home.component';
 import { ProjectsComponent } from './projects/projects.component';
 import { WebsitesComponent } from './websites/websites.component';
@@ -20,8 +19,16 @@ const routes: Routes = [
   {path: 'changes/add-feature', component: FeatureAddComponent}
 ];
 
+const routerOptions: ExtraOptions = {
+  useHash: false,
+  anchorScrolling: 'enabled',
+  onSameUrlNavigation: 'reload'
+}; 
+
+
+
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [RouterModule.forRoot(routes, routerOptions)],
   exports: [RouterModule]
 })
 export class AppRoutingModule { }
